# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Soko Moto - Dead Hours Flash Deal Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List deals
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, live, filling, sold_out, expired, completed]
        - name: category
          in: query
          schema:
            type: string
            enum: [restaurant, spa, fitness, bar, experience]
        - name: venueId
          in: query
          schema:
            type: integer
        - name: neighborhood
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a new flash deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealBody"
      responses:
        "201":
          description: Deal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get a deal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    patch:
      operationId: updateDeal
      tags: [deals]
      summary: Update deal status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDealBody"
      responses:
        "200":
          description: Updated deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"

  # ─── Bookings ───────────────────────────────────────────────────────────────
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: dealId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled, no_show]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a deal slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status (confirm, cancel, complete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ─── Users ──────────────────────────────────────────────────────────────────
...
  /ratings:
    get:
      operationId: listRatings
      tags: [ratings]
      summary: List ratings
      parameters:
        - name: venueId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rating"
    post:
      operationId: createRating
      tags: [ratings]
      summary: Submit a rating for a completed booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRatingBody"
      responses:
        "201":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"

  # ─── Standing Deals ─────────────────────────────────────────────────────────
  /standing-deals:
    get:
      operationId: listStandingDeals
      tags: [standing-deals]
      summary: List standing deals
      parameters:
        - name: venueId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of standing deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StandingDeal"
    post:
      operationId: createStandingDeal
      tags: [standing-deals]
      summary: Create a recurring standing deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStandingDealBody"
      responses:
        "201":
          description: Standing deal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandingDeal"
  /standing-deals/{id}:
    patch:
      operationId: updateStandingDeal
      tags: [standing-deals]
      summary: Update or pause a standing deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStandingDealBody"
      responses:
        "200":
          description: Updated standing deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandingDeal"
    delete:
      operationId: deleteStandingDeal
      tags: [standing-deals]
      summary: Delete a standing deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Feed (wow endpoints) ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Deal:
      type: object
      properties:
        id:
          type: integer
        venueId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        discountPercent:
          type: integer
        originalPrice:
          type: number
          nullable: true
        dealPrice:
          type: number
        totalSlots:
          type: integer
        bookedSlots:
          type: integer
        availableSlots:
          type: integer
        status:
          type: string
          enum: [draft, live, filling, sold_out, expired, completed]
        category:
          type: string
          enum: [restaurant, spa, fitness, bar, experience]
        dealType:
          type: string
          enum: [lunch, dinner, spa, wellness, fitness, drinks, experience]
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        isStandingDeal:
          type: boolean
        standingDealId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - venueId
        - title
        - discountPercent
        - dealPrice
        - totalSlots
        - bookedSlots
        - availableSlots
        - status
        - category
        - dealType
        - validFrom
        - validUntil
        - createdAt
...
    CreateDealBody:
      type: object
      properties:
        venueId:
          type: integer
        title:
          type: string
        description:
          type: string
        discountPercent:
          type: integer
        originalPrice:
          type: number
        dealPrice:
          type: number
        totalSlots:
          type: integer
        category:
          type: string
          enum: [restaurant, spa, fitness, bar, experience]
        dealType:
          type: string
          enum: [lunch, dinner, spa, wellness, fitness, drinks, experience]
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
      required:
        - venueId
        - title
        - discountPercent
        - dealPrice
        - totalSlots
        - category
        - dealType
        - validFrom
        - validUntil
    UpdateDealBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, live, filling, sold_out, expired, completed]
        title:
          type: string
        description:
          type: string
        totalSlots:
          type: integer

    # ─── Booking ─────────────────────────────────────────────────────────────
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        dealId:
          type: integer
        covers:
          type: integer
        totalPaid:
          type: number
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, no_show]
        bookingReference:
          type: string
        paymentMethod:
          type: string
          enum: [mpesa, card, corporate]
        paymentReference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        deal:
          $ref: "#/components/schemas/Deal"
      required:
        - id
        - userId
        - dealId
        - covers
        - totalPaid
        - status
        - bookingReference
        - paymentMethod
        - createdAt
    CreateBookingBody:
      type: object
      properties:
        userId:
          type: integer
        dealId:
          type: integer
        covers:
          type: integer
        paymentMethod:
          type: string
          enum: [mpesa, card, corporate]
        paymentReference:
          type: string
        notes:
          type: string
      required:
        - userId
        - dealId
        - covers
        - paymentMethod
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, no_show]
        paymentReference:
          type: string

    # ─── User ────────────────────────────────────────────────────────────────
...
    Rating:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        venueId:
          type: integer
        bookingId:
          type: integer
        score:
          type: integer
          minimum: 1
          maximum: 5
        review:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - venueId
        - bookingId
        - score
        - createdAt
    CreateRatingBody:
      type: object
      properties:
        userId:
          type: integer
        venueId:
          type: integer
        bookingId:
          type: integer
        score:
          type: integer
          minimum: 1
          maximum: 5
        review:
          type: string
      required:
        - userId
        - venueId
        - bookingId
        - score

    # ─── Standing Deal ───────────────────────────────────────────────────────
    StandingDeal:
      type: object
      properties:
        id:
          type: integer
        venueId:
          type: integer
        title:
          type: string
        discountPercent:
          type: integer
        dealPrice:
          type: number
        totalSlots:
          type: integer
        category:
          type: string
          enum: [restaurant, spa, fitness, bar, experience]
        dealType:
          type: string
          enum: [lunch, dinner, spa, wellness, fitness, drinks, experience]
        daysOfWeek:
          type: array
          items:
            type: integer
          description: "0=Sun, 1=Mon ... 6=Sat"
        startHour:
          type: integer
        endHour:
          type: integer
        activationThreshold:
          type: integer
          description: "Activate if bookings below this count by trigger hour"
        triggerHour:
          type: integer
          description: "Hour of day to check threshold and auto-activate"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - venueId
        - title
        - discountPercent
        - dealPrice
        - totalSlots
        - category
        - dealType
        - daysOfWeek
        - startHour
        - endHour
        - activationThreshold
        - triggerHour
        - isActive
        - createdAt
    CreateStandingDealBody:
      type: object
      properties:
        venueId:
          type: integer
        title:
          type: string
        discountPercent:
          type: integer
        dealPrice:
          type: number
        totalSlots:
          type: integer
        category:
          type: string
          enum: [restaurant, spa, fitness, bar, experience]
        dealType:
          type: string
          enum: [lunch, dinner, spa, wellness, fitness, drinks, experience]
        daysOfWeek:
          type: array
          items:
            type: integer
        startHour:
          type: integer
        endHour:
          type: integer
        activationThreshold:
          type: integer
        triggerHour:
          type: integer
      required:
        - venueId
        - title
        - discountPercent
        - dealPrice
        - totalSlots
        - category
        - dealType
        - daysOfWeek
        - startHour
        - endHour
        - activationThreshold
        - triggerHour
    UpdateStandingDealBody:
      type: object
      properties:
        isActive:
          type: boolean
        discountPercent:
          type: integer
        totalSlots:
          type: integer
        activationThreshold:
          type: integer

    # ─── Feed / Wow ──────────────────────────────────────────────────────────
...